requirements.txt
Flask==2.3.3
bootstrap-flask
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary
flask-sqlalchemy
